Split logic across BLoC and service files
