+++ lib/main.dart
@@ -1,143 +1,11 @@
+import 'package:currency_converter/screens/main.screen.dart';
 import 'package:flutter/material.dart';
-import 'package:http/http.dart' as http;
-import 'dart:async';
-import 'dart:convert';
 
 const requestURL = "https://api.hgbrasil.com/finance?format=json&key=bb1542a6";
 
-void main() async {
+void main() {
   runApp(MaterialApp(
-    home: new HomeWidget(),
+    home: MainScreen(),
     theme: ThemeData(hintColor: Colors.amber, primaryColor: Colors.white),
   ));
-}
-
-class HomeWidget extends StatefulWidget {
-  @override
-  _HomeWidgetState createState() => _HomeWidgetState();
-}
-
-class _HomeWidgetState extends State<HomeWidget> {
-  double dolar, euro;
-
-  final realController = TextEditingController();
-  final dolarController = TextEditingController();
-  final euroController = TextEditingController();
-
-  void _clearAll() {
-    realController.text = "";
-    dolarController.text = "";
-    euroController.text = "";
-  }
-
-  void _realChanged(String text) {
-    if (text.isEmpty) {
-      _clearAll();
-      return;
-    }
-
-    double real = double.parse(text);
-    dolarController.text = (real / dolar).toStringAsFixed(2);
-    euroController.text = (real / euro).toStringAsFixed(2);
-  }
-
-  void _dolarChanged(String text) {
-    if (text.isEmpty) {
-      _clearAll();
-      return;
-    }
-
-    double dolar = double.parse(text);
-    realController.text = (dolar * this.dolar).toStringAsFixed(2);
-    euroController.text = (dolar * this.dolar / euro).toStringAsFixed(2);
-  }
-
-  void _euroChanged(String text) {
-    if (text.isEmpty) {
-      _clearAll();
-      return;
-    }
-
-    double euro = double.parse(text);
-    realController.text = (euro * this.euro).toStringAsFixed(2);
-    dolarController.text = (euro * this.euro / dolar).toStringAsFixed(2);
-  }
-
-  Widget build(BuildContext context) {
-    return Scaffold(
-      backgroundColor: Colors.black,
-      appBar: AppBar(
-        title: Text("\$ Conversor \$"),
-        backgroundColor: Colors.amber,
-        centerTitle: true,
-      ),
-      body: FutureBuilder<Map>(
-        future: getData(),
-        builder: (context, snapshot) {
-          switch (snapshot.connectionState) {
-            case ConnectionState.none:
-            case ConnectionState.waiting:
-              return Center(
-                child: Text(
-                  "Carregando dados...",
-                  style: TextStyle(color: Colors.amber),
-                  textAlign: TextAlign.center,
-                ),
-              );
-            default:
-              if (snapshot.hasError) {
-                return Center(
-                  child: Text(
-                    "Erro ao carregar dados :(",
-                    style: TextStyle(color: Colors.amber),
-                    textAlign: TextAlign.center,
-                  ),
-                );
-              } else {
-                dolar = snapshot.data["USD"]["buy"];
-                euro = snapshot.data["EUR"]["buy"];
-
-                return SingleChildScrollView(
-                  padding: EdgeInsets.all(10.0),
-                  child: Column(
-                    crossAxisAlignment: CrossAxisAlignment.stretch,
-                    children: <Widget>[
-                      Icon(Icons.monetization_on,
-                          size: 150.0, color: Colors.amber),
-                      buildTextField(
-                          "Reais", "R\$", realController, _realChanged),
-                      Divider(),
-                      buildTextField(
-                          "Dólares", "US\$", dolarController, _dolarChanged),
-                      Divider(),
-                      buildTextField("Euros", "€", euroController, _euroChanged)
-                    ],
-                  ),
-                );
-              }
-          }
-        },
-      ),
-    );
-  }
-}
-
-Widget buildTextField(
-    String label, String prefix, TextEditingController c, Function f) {
-  return TextField(
-    controller: c,
-    decoration: InputDecoration(
-        labelText: label,
-        labelStyle: TextStyle(color: Colors.amber),
-        border: OutlineInputBorder(),
-        prefixText: prefix),
-    style: TextStyle(color: Colors.amber, fontSize: 25.0),
-    onChanged: f,
-    keyboardType: TextInputType.numberWithOptions(decimal: true),
-  );
-}
-
-Future<Map> getData() async {
-  http.Response response = await http.get(requestURL);
-  return json.decode(response.body)['results']['currencies'];
 }
