+++ pubspec.yaml
@@ -24,6 +24,7 @@ dependencies:
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^0.1.2
   http: ^0.12.0+2
+  rxdart: ^0.22.0
   
 dev_dependencies:
   flutter_test:
